Simplify exchange creator declaration
