+++ src/main/java/com/github/fridujo/rabbitmq/mock/exchange/TypedMockExchangeCreator.java
@@ -1,0 +1,6 @@
+package com.github.fridujo.rabbitmq.mock.exchange;
+
+public interface TypedMockExchangeCreator extends MockExchangeCreator {
+
+    String getType();
+}
